Refactor hearing test welcome page structure by moving tabs to separate files
